requirements.txt

certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
loguru==0.7.2
pydub==0.25.1
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot-calendar==1.0.5
requests==2.32.3
six==1.16.0
soupsieve==2.6
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.2